Moved fakes into fakes folder
