+++ fakes/twitterfake.js
